Spatial filter customer

+++ Application/DTO/CustomerListFilterDto.cs
@@ -6,7 +6,7 @@ namespace Application.DTO
     {
         public double Latitude { get; set; }
         public double Longitude { get; set; }
-        public long Distance { get; set; }
+        public double Distance { get; set; }
         public double Rating { get; set; }
         public IEnumerable<string> Tags { get; set; }
         public string Type { get; set; } // Um especificador mais geral do que mais representa o estabelecimento
